libs/services/ocr_engine.py
"""
OCR Engine using Google ML Kit Text Recognition.
Requires Android platform and pyjnius for Java bridge.
"""

class OCREngine:
    """
    Optical Character Recognition engine using Google ML Kit.
    Extracts text from images captured via camera or selected from gallery.
    """
    
    def __init__(self):
        self.recognizer = None
        self.is_loaded = False
        self.platform = self._detect_platform()
    
    def _detect_platform(self) -> str:
        """Detect if running on Android."""
        try:
            from jnius import autoclass
            return "android"
        except ImportError:
            return "other"
    
    def load(self) -> bool:
        """Initialize ML Kit Text Recognition."""
        if self.platform != "android":
            print("OCR is only available on Android")
            return False
        
        try:
            from jnius import autoclass
            
            # Import ML Kit classes
            TextRecognition = autoclass('com.google.mlkit.vision.text.TextRecognition')
            TextRecognizerOptions = autoclass('com.google.mlkit.vision.text.latin.TextRecognizerOptions')
            
            # Create recognizer
            options = TextRecognizerOptions.Builder().build()
            self.recognizer = TextRecognition.getClient(options)
            
            self.is_loaded = True
            return True
            
        except Exception as e:
            print(f"Failed to load ML Kit OCR: {e}")
            return False
    
    def extract_text_from_image(self, image_path: str) -> str:
        """
        Extract text from an image file.
        
        Args:
            image_path: Path to image file
            
        Returns:
            Extracted text
        """
        if not self.is_loaded:
            return "Error: OCR engine not loaded"
        
        if self.platform != "android":
            return "Error: OCR only available on Android"
        
        try:
            from jnius import autoclass
            
            # Import Android classes
            InputImage = autoclass('com.google.mlkit.vision.common.InputImage')
            File = autoclass('java.io.File')
            
            # Create InputImage from file
            image_file = File(image_path)
            image = InputImage.fromFilePath(None, image_file)  # Context can be None for file paths
            
            # Process image (synchronous for simplicity)
            task = self.recognizer.process(image)
            
            # Wait for result (this is pseudo-code - actual implementation needs proper async handling)
            # In production, use task.addOnSuccessListener() and callbacks
            result = task.getResult()
            
            if result:
                return result.getText()
            return ""
            
        except Exception as e:
            return f"Error extracting text: {e}"
    
    def extract_text_from_camera(self, callback):
        """
        Capture image from camera and extract text.
        
        Args:
            callback: Function to call with extracted text
        """
        # This requires camera integration with Kivy/Android
        # Placeholder implementation
        if self.platform != "android":
            callback("Error: Camera only available on Android")
            return
        
        # TODO: Implement camera capture using plyer or android API
        # 1. Request camera permission
        # 2. Open camera activity
        # 3. Capture image
        # 4. Pass to extract_text_from_image
        # 5. Call callback with result
        
        callback("Camera capture not yet implemented")
    
    def unload(self) -> None:
        """Unload the OCR engine."""
        if self.recognizer and self.platform == "android":
            try:
                self.recognizer.close()
            except:
                pass
        self.recognizer = None
        self.is_loaded = False


# Stub implementation for non-Android platforms
class OCREngineStub:
    """
    Stub OCR engine for testing on non-Android platforms.
    """
    
    def __init__(self):
        self.is_loaded = False
    
    def load(self) -> bool:
        print("Using OCR stub (Android required for real OCR)")
        self.is_loaded = True
        return True
    
    def extract_text_from_image(self, image_path: str) -> str:
        return "Mock OCR text: This is sample extracted text from an image."
    
    def extract_text_from_camera(self, callback):
        callback("Mock OCR text from camera")
    
    def unload(self) -> None:
        self.is_loaded = False
